+++ SDL/TrackExtensions.cuh
@@ -30,8 +30,8 @@ namespace SDL
     {
         short* constituentTCTypes;
         unsigned int* constituentTCIndices;
-        unsigned int* nLayerOverlaps;
-        unsigned int* nHitOverlaps;
+        uint8_t* nLayerOverlaps;
+        uint8_t* nHitOverlaps;
         unsigned int* nTrackExtensions; //overall counter!
         FPX* rPhiChiSquared;
         FPX* rzChiSquared;
